llm-agent/requirements.txt
fastapi[all]
uvicorn
requests
PyJWT
passlib
aiohttp
python-jose
pydantic
python-dotenv 